Feat: Add Follow Recommend

See merge request s07-webmobile1-sub2/S07P12A603!227

+++ backend/src/main/java/com/a603/youlangme/dto/follow/FollowCandidate.java
@@ -5,9 +5,9 @@ import lombok.Getter;
 @Getter
 public class FollowCandidate {
     private Long id;
-    private Float sim;
+    private Double sim;
 
-    static public FollowCandidate of(Long id, Float sim) {
+    static public FollowCandidate of(Long id, Double sim) {
         FollowCandidate candidate = new FollowCandidate();
         candidate.id = id;
         candidate.sim = sim;
